public-safety-service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.53
pydantic==1.8.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.0
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
